+++ src/shared/model/config/IComponentsIds.ts
@@ -2,6 +2,6 @@ export interface IComponentsIds {
   animatedCard?: string;
   cardNumber: string;
   expirationDate: string;
-  notificationFrame: string;
+  notificationFrame?: string;
   securityCode: string;
 }